Rename meta picker view modules
