+++ app/src/features/meta-picker/ui/screens/meta-picker/index.tsx
@@ -3,9 +3,9 @@ import { View, FlatList, ListRenderItem } from 'react-native'
 
 import { NavigationScreen } from 'src/ui/types'
 import Logic, { Props, State, Event } from './logic'
-import MetaPicker from '../../components/meta-picker'
-import MetaPickerEntry from '../../components/meta-picker-entry'
-import MetaPickerEmptyRow from '../../components/meta-picker-empty'
+import MetaPicker from '../../components/picker'
+import MetaPickerEntry from '../../components/picker-entry'
+import MetaPickerEmptyRow from '../../components/picker-entry-empty'
 import SuggestInput from '../../components/suggest-input'
 import * as selectors from './selectors'
 import { MetaTypeShape, MetaTypeName } from 'src/features/meta-picker/types'
